Hangman
This is a hangman game that we all know. Only difference is it is with coding. This is a beginner level project. But for all the people who do not know what it is here is the how it is played.

When you run the project computer decides a word but it does not shown you.
When you guess a letter and if it is in the decided word, it is shown you where is the letter in the word. On the other hand, if it not in the word. then you get close to create a hangman.
You continue this proccess up to 6 times of your fail.
If you could not guess the word and hangman is created, it means that you lost.
If you guessed the word before hangman created, it means you won!

No matter what happens I am on your side :)